feat(knit): drop YAML output gate; all formats preview as HTML
- yaml-frontmatter.ts: drop 'non-html-format' from BlockerKind union
- knit-commands.ts: remove the isSupportedHtmlFormat gate + buildNonHtmlFormatBlocker
- Rename knit-html-only.test.ts -> knit-yaml-output-ignored.test.ts;
  rewrite assertions to verify every output format launches knit
- Move bun tests from editors/vscode/src/knit/*.test.ts to tests/bun/
  (project convention; vscode tsconfig excludes them from typecheck)
